Fix the config import
