+++ cli/package.json
@@ -1,11 +1,11 @@
 {
   "name": "@openapi-codegen/cli",
-  "version": "0.0.1",
+  "version": "0.0.1-alpha.1",
   "description": "OpenAPI Codegen cli",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "bin": {
-    "openapi-codegen": "lib/index.js"
+    "openapi-codegen": "lib/cli.js"
   },
   "repository": {
     "url": "https://github.com/fabien0102/openapi-codegen",
@@ -20,7 +20,7 @@
   ],
   "license": "MIT",
   "scripts": {
-    "start": "node -r @swc-node/register src/index.ts",
+    "start": "node -r @swc-node/register src/cli.ts",
     "build": "tsc -p tsconfig.package.json"
   },
   "files": [
